requirements.txt
fastapi==0.104.1
openai==0.28.1
uvicorn==0.24.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
langchain == 0.0.336
chromadb==0.4.17
openai==1.3.3
pydantic-settings==2.1.0
sentence-transformers
yt-dlp
pytube